@@ -12,16 +12,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.2.4",
+    "@angular/animations": "^18.2.10",
     "@angular/cdk": "^17.0.0",
-    "@angular/common": "^17.2.4",
-    "@angular/compiler": "^17.2.4",
-    "@angular/core": "^17.2.4",
-    "@angular/forms": "^17.2.4",
-    "@angular/platform-browser": "^17.2.4",
-    "@angular/platform-browser-dynamic": "^17.2.4",
-    "@angular/router": "^17.2.4",
-    "@angular/service-worker": "^17.2.4",
+    "@angular/common": "^18.2.10",
+    "@angular/compiler": "^18.2.10",
+    "@angular/core": "^18.2.10",
+    "@angular/forms": "^18.2.10",
+    "@angular/platform-browser": "^18.2.10",
+    "@angular/platform-browser-dynamic": "^18.2.10",
+    "@angular/router": "^18.2.10",
+    "@angular/service-worker": "^18.2.10",
     "ng": "^0.0.0",
     "ol": "7.1.0",
     "primeflex": "^2.0.0",
@@ -29,14 +29,14 @@
     "primeng": "^17.0.0",
     "rxjs": "^7.4",
     "tslib": "^2.3.0",
-    "zone.js": "~0.14.4"
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.2.3",
-    "@angular/cli": "^17.2.3",
-    "@angular/compiler-cli": "^17.2.4",
-    "@angular/language-service": "^17.2.4",
-    "@angular/localize": "^17.2.4",
+    "@angular-devkit/build-angular": "^18.2.11",
+    "@angular/cli": "^18.2.11",
+    "@angular/compiler-cli": "^18.2.10",
+    "@angular/language-service": "^18.2.10",
+    "@angular/localize": "^18.2.10",
     "@types/jasmine": "~3.10.3",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^20.0.0",
@@ -50,6 +50,6 @@
     "karma-jasmine-html-reporter": "^1.5.0",
     "karma-junit-reporter": "^2.0.1",
     "tslint": "~6.1.0",
-    "typescript": "~5.3.3"
+    "typescript": "~5.5.4"
   }
 }
